REST API
 - Use get_obeject_or_404
 - Apply REST API to project

+++ articles/views.py
@@ -1,20 +1,26 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Article
+
 
 # Create your views here.
 
-def new(request):
-    return render(request, 'articles/new.html')
-
+# def new(request):
+#     return render(request, 'articles/new.html')
 
 def create(request):
-    title = request.GET.get('title')
-    content = request.GET.get('content')
-    article = Article()
-    article.title = title
-    article.content = content
-    article.save()
-    return redirect(f'/articles/{article.pk}')
+    # 아니라면 POST 일경우 사용자 데이터 받아서 article 생성
+    if request.method == 'POST':
+        title = request.POST.get('title')
+        content = request.POST.get('content')
+        article = Article()
+        article.title = title
+        article.content = content
+        article.save()
+        return redirect('articles:detail', article.pk)
+    #만약 GET 요청으로 들어오면 HTML 페이지 rendering
+    else:
+        return render(request, 'articles/create.html')
+        
 
 
 def index(request):
@@ -25,13 +31,28 @@ def index(request):
 
 
 def detail(request, article_pk):
-    article = Article.objects.get(pk=article_pk)
+    article = get_object_or_404(Article, pk=article_pk)
     context = {'article': article}
     return render(request, 'articles/detail.html', context)
 
 
 def delete(request, article_pk):
-    article = Article.objects.get(pk=article_pk)
-    article.delete()
-    return redirect('/articles/')
+    article = get_object_or_404(Article, pk=article_pk)
+    if request.method == 'POST':
+        article.delete()
+        return redirect('articles:index')
+    else:
+        return redirect('articles:detail', article.pk) 
 
+
+def update(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+    if request.method == 'POST':
+        title = request.POST.get('title')
+        content = request.POST.get('content')
+        article.title = title
+        article.content = content
+        article.save()
+        return redirect('articles:index')
+    else:
+        return render(request, 'articles/update.html', {'article': article})
